Virtual DHCP: add feature to define reserved/static IP addresses
A static IP address is assigned to a client when the specified MAC address and hostname match. A value is only checked against when it's not empty.

When both the MAC address and hostname are empty or not defined in the configuration file, the IP address is excluded from the DHCP pool and it can only be assigned on the client.